ESP32 Up Time
Displays the Esp32 Chip Ip address (In this case  a FreeNove WRoom)    the chip hosts a simple website at the ip address which displays current up time of chip as well as random chip data monitoring.    This was intended to see how long a battery would last connected to the esp32.